requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
asgiref==3.8.1
    # via
    #   django
    #   django-htmx
asttokens==2.4.1
    # via stack-data
beautifulsoup4==4.12.3
    # via o365
black==24.10.0
    # via -r requirements.in
boto3==1.35.46
    # via -r requirements.in
botocore==1.35.46
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   djlint
colorama==0.4.6
    # via djlint
coverage==7.6.4
    # via -r requirements.in
cryptography==43.0.3
    # via pyjwt
cssbeautifier==1.15.1
    # via djlint
decorator==5.1.1
    # via ipython
djade==1.3.1
    # via -r requirements.in
django==5.1.2
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-auto-prefetch
    #   django-extensions
    #   django-htmx
    #   django-markdownx
    #   django-modeltranslation
    #   django-rosetta
    #   django-storages
django-allauth==65.0.2
    # via -r requirements.in
django-auto-prefetch==1.10.0
    # via -r requirements.in
django-cleanup==9.0.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-htmx==1.19.0
    # via -r requirements.in
django-markdownx==4.0.7
    # via -r requirements.in
django-modeltranslation==0.19.10
    # via -r requirements.in
django-o365mail==1.1.0
    # via -r requirements.in
django-rosetta==0.10.1
    # via -r requirements.in
django-storages==1.14.4
    # via -r requirements.in
django-upgrade==1.22.1
    # via -r requirements.in
djlint==1.35.2
    # via -r requirements.in
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
executing==2.1.0
    # via stack-data
factory-boy==3.3.1
    # via -r requirements.in
faker==30.8.0
    # via factory-boy
gunicorn==23.0.0
    # via -r requirements.in
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
huey==2.5.2
    # via -r requirements.in
idna==3.10
    # via requests
ipython==8.28.0
    # via -r requirements.in
isort==5.13.2
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.25
    # via djlint
markdown==3.7
    # via django-markdownx
matplotlib-inline==0.1.7
    # via ipython
mypy-extensions==1.0.0
    # via black
o365==2.0.37
    # via django-o365mail
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   black
    #   gunicorn
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   djlint
pexpect==4.9.0
    # via ipython
pillow==11.0.0
    # via
    #   -r requirements.in
    #   django-markdownx
platformdirs==4.3.6
    # via black
polib==1.2.0
    # via django-rosetta
prompt-toolkit==3.0.48
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyclean==3.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via ipython
pyjwt==2.9.0
    # via django-allauth
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   faker
    #   o365
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via djlint
redis==5.1.1
    # via -r requirements.in
regex==2024.9.11
    # via djlint
requests==2.32.3
    # via
    #   django-allauth
    #   django-rosetta
    #   o365
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via
    #   django-allauth
    #   o365
ruff==0.7.0
    # via -r requirements.in
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via
    #   asttokens
    #   cssbeautifier
    #   jsbeautifier
    #   python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
tokenize-rt==6.1.0
    # via django-upgrade
tqdm==4.66.5
    # via djlint
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via faker
tzdata==2024.2
    # via o365
tzlocal==5.2
    # via o365
urllib3==2.2.3
    # via
    #   botocore
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
